Added edit Snippet properties form (first part)
In this patch I have added the possibility of editing the properties of a snippet.
The list of Snippets codes for a snippet and the creation of a new one are not yet done. With the form added only the title and description of a existing snippet can be changed.

The old button edit from 'List of Snippets' form has now the label 'Rename', as the old code allows only change the title.