Corrected 1-N relationship between Snippet and SnippetCode in Hibernate Mapping file
If there is a snippet_id column in snippet table you have to delete it.

alter table snippet drop column snippet_id;

Also if there is a constraint relating id column from snippet_code with id column of snippet it has to be deleted also. This errors in database structure are due to incorrect mapping