Added relationship 1-N between Label and Snippet. Corrected relationship 1-N between User and Label entities (2nd try)
The database structure has changed (when correcting 1-N relationships the foreign keys in some tables are now removed or added). Please backup your data to insert it after the recreating the database.
A SQL script called example_data.sql is included to make it easier to fill the database with test data if you don't have any already created or don't want to recreate it